requirements.txt
aniso8601==3.0.0
astroid==1.6.3
atomicwrites==1.1.5
attrs==18.1.0
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
coveralls==1.3.0
docopt==0.6.2
flasgger==0.8.3
Flask==0.12.2
Flask-Login==0.4.1
Flask-RESTful==0.3.6
Flask-Testing==0.7.1
Flask-WTF==0.14.2
gunicorn==19.8.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.2.0
nose==1.3.7
pep8==1.7.1
pluggy==0.6.0
py==1.5.3
pycodestyle==2.4.0
PyJWT==1.6.1
pylint==1.8.4
pytest==3.6.1
pytest-cov==2.5.1
pytz==2018.4
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
